+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning].
 
 - Add option to explicitly state map file format. ([#46])
 - Add support for more extensions as MarkDown map file. ([#50])
+- Allow file extension of map files to be upper or mixed case. ([#55])
 
 ### Bug Fixes
 
@@ -73,3 +74,4 @@ Versioning].
 [#42]: https://github.com/ericcornelissen/wordrow/pull/42
 [#46]: https://github.com/ericcornelissen/wordrow/pull/46
 [#50]: https://github.com/ericcornelissen/wordrow/pull/50
+[#55]: https://github.com/ericcornelissen/wordrow/pull/55